+++ rdf_fdw--2.2--2.3.sql
@@ -1,1 +1,64 @@
-\echo "No changes"
+
+CREATE OR REPLACE FUNCTION sparql.hours(time)
+RETURNS int AS $$
+BEGIN
+  RETURN EXTRACT(hour FROM $1);
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;
+
+CREATE OR REPLACE FUNCTION sparql.seconds(rdfnode)
+RETURNS numeric AS $$
+BEGIN
+    DECLARE
+        dt text := sparql.datatype($1)::text;
+    BEGIN
+        IF dt = '<http://www.w3.org/2001/XMLSchema#time>' THEN
+            RETURN EXTRACT(second FROM sparql.lex($1)::time);
+        ELSE
+            RETURN EXTRACT(second FROM sparql.lex($1)::timestamp);
+        END IF;
+    END;
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;
+
+CREATE OR REPLACE FUNCTION sparql.minutes(rdfnode)
+RETURNS int AS $$
+BEGIN
+    DECLARE
+        dt text := sparql.datatype($1)::text;
+    BEGIN
+        IF dt = '<http://www.w3.org/2001/XMLSchema#time>' THEN
+            RETURN EXTRACT(minute FROM sparql.lex($1)::time);
+        ELSE
+            RETURN EXTRACT(minute FROM sparql.lex($1)::timestamp);
+        END IF;
+    END;
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;
+
+CREATE OR REPLACE FUNCTION sparql.hours(rdfnode)
+RETURNS int AS $$
+DECLARE
+    dt text := sparql.datatype($1)::text;
+BEGIN
+    IF dt = '<http://www.w3.org/2001/XMLSchema#time>' THEN
+        RETURN EXTRACT(hour FROM sparql.lex($1)::time);
+    ELSE
+        RETURN EXTRACT(hour FROM sparql.lex($1)::timestamp);
+    END IF;
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;
+
+CREATE OR REPLACE FUNCTION sparql.minutes(time)
+RETURNS int AS $$
+BEGIN
+  RETURN EXTRACT(minute FROM $1);
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;
+
+CREATE OR REPLACE FUNCTION sparql.seconds(time)
+RETURNS numeric AS $$
+BEGIN
+  RETURN EXTRACT(second FROM $1);
+END;
+$$ LANGUAGE plpgsql STABLE STRICT;